render with console log, state implementation
