+++ index.html
@@ -8,8 +8,8 @@
 </head>
 <body>
 <h1>Snake Game!</h1>
-<button id="start-game">Start the Game!</button>
-<canvas width="600" height="600" id="gameboard"></canvas>
+<canvas width="600" height="600" id="gameboard">
+</canvas>
 <script src ="./index.js"></script>
 </body>
 </html>
