+++ package.json
@@ -1,12 +1,20 @@
 {
   "name": "snakegame",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
-  "license": "ISC"
+  "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/NikolaCirisanovic/snakegame.git"
+  },
+  "bugs": {
+    "url": "https://github.com/NikolaCirisanovic/snakegame/issues"
+  },
+  "homepage": "https://github.com/NikolaCirisanovic/snakegame#readme",
+  "description": ""
 }
