+++ render.js
@@ -1,0 +1,6 @@
+const canvas = document.getElementById("gameboard");
+const ctx = canvas.getContext("2d");
+
+export function paint(state) {
+  console.log("state", state);
+}
